Add Mongoose models including their schemas to the server
The models have been created based on the collections which can be found in the database structure. The validation rules of each model beyond primary keys have
NOT yet been added.
This commit also includes a simple test that adds a quiznight to the database. The test checks if it has actually been added and removes the data once it completed.